% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt4r_addsessionid.R
\name{mt4r_addsessionid}
\alias{mt4r_addsessionid}
\title{Add session_id column}
\usage{
mt4r_addsessionid(dat)
}
\arguments{
\item{dat}{A data frame with the following columns: event_params.key,
event_date, event_params.value.int_value, row_id. The input data.frame is
created by the \code{\link[=mt4r_unnest]{mt4r_unnest()}} function.}
}
\value{
A data frame with the session_id column added.
}
\description{
This function adds a session_id column to the data.
The session_id is a unique identifier for each session.
}
\examples{
mockdata |>
  mt4r_unnest() |>
  mt4r_addsessionid()
}
