% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt4r_fixtime.R
\name{mt4r_fixtime}
\alias{mt4r_fixtime}
\title{Fix the timestamp column in the data}
\usage{
mt4r_fixtime(dat)
}
\arguments{
\item{dat}{A data frame with the following columns: event_timestamp,
event_params.value.int_value, event_params.value.double_value. The input
data.frame is created by the \code{\link[=mt4r_unnest]{mt4r_unnest()}} and optionally modified by the
\code{\link[=mt4r_addsessionid]{mt4r_addsessionid()}} function.}
}
\value{
A data frame with the \code{event_timestamp} column converted into a
datetime object (\code{event_timestamp2}).
}
\description{
The \code{event_timestamp} column in the data is a character vector, which
represents time passed since the unix epoch in microseconds. This function
converts the \code{event_timestamp} column into a datetime object.
}
\examples{
# To be added
}
