% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt4r_routelookup.R
\name{mt4r_routelookup}
\alias{mt4r_routelookup}
\title{Creating a lookup table for route starts}
\usage{
mt4r_routelookup(dat)
}
\arguments{
\item{dat}{A data frame with columns row_id, event_name, and event_timestamp2}
}
\value{
A data frame with columns row_id, has_route_started, and
route_started_number
}
\description{
This function creates a lookup table for route starts. It takes a data frame
and returns a data frame with a running number for each route start.
}
\examples{
mockdata |>
  mt4r_unnest() |>
  mt4r_fixtime() |>
  mt4r_addsessionid() |>
  mt4r_routelookup()
}
