% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt4r_routestarts.R
\name{mt4r_routestarts}
\alias{mt4r_routestarts}
\title{Add route name and route start numbers to data}
\usage{
mt4r_routestarts(dat, route_lookup, filter_zero_data = TRUE)
}
\arguments{
\item{dat}{A data frame with the following columns: row_id, session_id,
event_name, event_params.key, event_params.value.string_value. The input
data.frame is created by the \code{\link[=mt4r_unnest]{mt4r_unnest()}}, \code{\link[=mt4r_fixtime]{mt4r_fixtime()}}, and
\code{\link[=mt4r_addsessionid]{mt4r_addsessionid()}} functions.}

\item{route_lookup}{A lookup  data frame with the running route start
numbers. Is created by the \code{\link[=mt4r_routelookup]{mt4r_routelookup()}} function.}

\item{filter_zero_data}{A logical value. If TRUE, rows where a route has not
started are removed from the data.}
}
\value{
A data frame with the \code{route_started_number}, \code{route_id} and
\code{route_name} columns added to the input data.
}
\description{
This function adds a running route start number to the data. Additionally,
a \code{route_name} column is added to the data, to help easily connect route
starts to specific routes.
}
\examples{
dat <- mockdata |>
  mt4r_unnest() |>
  mt4r_addsessionid() |>
  mt4r_fixtime()

 lookup_dat <- dat |>
  mt4r_routelookup()

 dat |>
   mt4r_routestarts(lookup_dat)
}
