% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt4r_stoptagvals.R
\name{mt4r_stoptagvals}
\alias{mt4r_stoptagvals}
\title{Get tag values for each route stop}
\usage{
mt4r_stoptagvals(dat)
}
\arguments{
\item{dat}{Input data.frame containing the columns row_id, event_name,
event_params.key, event_params.value.int_value, and
event_params.value.double_value. The input data is created with the functions
\code{\link[=mt4r_unnest]{mt4r_unnest()}}, \code{\link[=mt4r_addsessionid]{mt4r_addsessionid()}}, and \code{\link[=mt4r_fixtime]{mt4r_fixtime()}}.}
}
\value{
A data.frame with the columns row_id, session_id, event_name,
route_id, engagement_time (in seconds), and route_pois_finished (percentage).
}
\description{
Gets the numeric values for route_id, engagement_time, and route_pois_finished
related to each unique route_stopped event.
}
\examples{
mockdata |>
  mt4r_unnest() |>
  mt4r_addsessionid() |>
  mt4r_fixtime() |>
  mt4r_stoptagvals()
}
