% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt4r_unnest.R
\name{mt4r_unnest}
\alias{mt4r_unnest}
\title{Unnesting json data}
\usage{
mt4r_unnest(dat)
}
\arguments{
\item{dat}{A data frame with columns \code{event_params}, \code{event_date}, and
\code{user_properties}. Created by reading in the exported data from Firebase.}
}
\value{
An unnested data frame with \code{user_properties} column removed.
}
\description{
Data exported from Firebase is in json format. This function unnests the
json data.
}
\examples{
# To be added
}
