% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exertime.R
\name{plot_exertime}
\alias{plot_exertime}
\title{Plot time spent on app}
\usage{
plot_exertime(
  dat,
  time_units = c("mins", "secs", "hours", "auto"),
  col_fill = "#99ff99",
  col_color = "gray",
  alpha_val = 0.9
)
}
\arguments{
\item{dat}{A data frame with columns \code{session_id} and \code{event_timestamp2}}

\item{time_units}{a character string specifying the time units to use
c("mins", "secs", "hours","auto")}

\item{col_fill}{a character string specifying the fill color}

\item{col_color}{a character string specifying the color of the histogram bar
outlines}

\item{alpha_val}{transparency of the histogram bars. Numeric value between 0
and 1}
}
\value{
a ggplot object
}
\description{
This convenience function plots the time spent on the app per session id.
}
\examples{
mockdata |>
  mt4r_unnest() |>
  mt4r_fixtime() |>
  mt4r_addsessionid() |>
  plot_exertime()
}
