% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_popular.R
\name{plot_popular}
\alias{plot_popular}
\title{Plot popular routes}
\usage{
plot_popular(
  dat,
  x_var,
  y_var,
  col_fill = "#99ff99",
  col_color = "gray",
  alpha_val = 0.9
)
}
\arguments{
\item{dat}{a data frame with the columns named after the values set for
\code{x_var} and \code{y_var}}

\item{x_var}{a column name in \code{dat} that will be used as the x-axis}

\item{y_var}{a column name in \code{dat} that will be used as the y-axis}

\item{col_fill}{a character string representing the fill color of the bars}

\item{col_color}{a character string representing the color of the bar edges}

\item{alpha_val}{a numeric value between 0 and 1 representing the
transparency of the bars}
}
\value{
a ggplot object
}
\description{
Create a bar plot of the most popular routes.
}
\examples{
data.frame(n_starts = c(1, 2, 3), route = c("A", "B", "C")) |>
  plot_popular(x_var = n_starts, y_var = route)
}
